torch_segformer/runpod_requirements.txt
glob2
plotly
scikit-learn
transformers
transformers[torch]
torchvision
opencv-python
albumentations

pip install --upgrade kaleido
pip install -U mlflow --ignore-installed